requirements.txt
python-dotenv
fastapi[standard]
